requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==4.4.7
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
html5lib==1.1
idna==3.2
kombu==4.6.11
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
pydot==1.4.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyphen==0.11.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.6
uWSGI==2.0.19.1
vine==1.3.0
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.1
XlsxWriter==1.4.4
